refactor: extract shared constants and utilities to shared/pem_common.py
Eliminates ~115 lines of duplicated code across cluster-link, mirror-topics,
and connect-replicator. SN_SOURCE_CLUSTERS, SN_BROKERS_PER_CLUSTER,
CONFLUENT_INSTALL, check_confluent_cli, check_auth, and load_pem_files now
live in shared/pem_common.py; each tool imports via sys.path.

Also fixes mirror_topics.py: default --config path is now resolved relative
to the script file (not CWD), and a missing warning is added when the mirror
list JSON fails to parse.
