+++ cluster-link/create_link.py
@@ -12,34 +12,26 @@ import argparse
 import configparser
 import json
 import os
-import shutil
 import subprocess
 import sys
 import tempfile
 import time
 
+_ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+if _ROOT not in sys.path:
+    sys.path.insert(0, _ROOT)
+from shared.pem_common import (
+    SN_SOURCE_CLUSTERS, SN_BROKERS_PER_CLUSTER, CONFLUENT_INSTALL,
+    check_confluent_cli, check_auth, load_pem_files,
+)
+
 
 REQUIRED_KEYS = ("environment_id", "cluster_id", "link_name")
-
-# ServiceNow active-active: base port for each source cluster, 4 brokers each.
-# These are the defaults; both can be overridden in link.conf.
-SN_SOURCE_CLUSTERS = [4100, 4200]
-SN_BROKERS_PER_CLUSTER = 4
 
 
 def sn_bootstrap(host: str, base_port: int, brokers_per_cluster: int = SN_BROKERS_PER_CLUSTER) -> str:
     """Build a bootstrap string for one ServiceNow source cluster."""
     return ",".join(f"{host}:{base_port + i}" for i in range(brokers_per_cluster))
-
-CONFLUENT_INSTALL = """\
-Confluent CLI not found. Install it:
-
-  Windows : winget install Confluent.ConfluentCLI
-  macOS   : brew install confluentinc/tap/confluent-cli
-  Linux   : See https://docs.confluent.io/confluent-cli/current/install.html
-
-Then authenticate: confluent login
-"""
 
 
 def load_config(path: str) -> dict:
@@ -74,52 +66,6 @@ def load_config(path: str) -> dict:
     bpc_raw = result.get("brokers_per_cluster", "")
     result["brokers_per_cluster"] = int(bpc_raw) if bpc_raw else SN_BROKERS_PER_CLUSTER
     return result
-
-
-def check_confluent_cli() -> None:
-    """Exit 1 with install instructions if 'confluent' is not in PATH."""
-    if shutil.which("confluent") is None:
-        print(CONFLUENT_INSTALL)
-        sys.exit(1)
-
-
-def check_auth(environment_id: str, cluster_id: str) -> None:
-    """Verify Confluent CLI auth by describing the destination cluster."""
-    result = subprocess.run(
-        [
-            "confluent", "kafka", "cluster", "describe",
-            cluster_id,
-            "--environment", environment_id,
-        ],
-        capture_output=True,
-        text=True,
-    )
-    if result.returncode != 0:
-        print(
-            "Error: Not authenticated or cluster not accessible. Run: confluent login",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-
-def load_pem_files(pem_dir: str):
-    """Load ca.pem, client-cert.pem, client-key.pem; exit 1 if any are missing."""
-    files = {
-        "ca.pem": None,
-        "client-cert.pem": None,
-        "client-key.pem": None,
-    }
-    for name in files:
-        path = os.path.join(pem_dir, name)
-        if not os.path.exists(path):
-            print(
-                f"Error: {name} not found in {pem_dir}. Run: python extract_pem.py first",
-                file=sys.stderr,
-            )
-            sys.exit(1)
-        with open(path, "rb") as fh:
-            files[name] = fh.read()
-    return files["ca.pem"], files["client-cert.pem"], files["client-key.pem"]
 
 
 def _inline(pem_bytes: bytes, literal_newlines: bool = False) -> str:
